+++ src/api/autorun.ts
@@ -1,4 +1,4 @@
-import { Lambda, getNextId, invariant, EMPTY_OBJECT, deprecated } from "../utils/utils"
+import { Lambda, getNextId, invariant, EMPTY_OBJECT } from "../utils/utils"
 import { Reaction, IReactionPublic, IReactionDisposer } from "../core/reaction"
 import { action, isAction } from "./action"
 import { IEqualsComparer, comparer } from "../utils/comparer"
@@ -87,12 +87,6 @@ export function reaction<T>(
     effect: (arg: T, r: IReactionPublic) => void,
     opts: IReactionOptions = EMPTY_OBJECT
 ): IReactionDisposer {
-    if (typeof opts === "boolean") {
-        opts = { fireImmediately: opts }
-        deprecated(
-            `Using fireImmediately as argument is deprecated. Use '{ fireImmediately: true }' instead`
-        )
-    }
     if (process.env.NODE_ENV !== "production") {
         invariant(
             typeof expression === "function",